+++ R/data.R
@@ -10,11 +10,11 @@
 #' IslamicArt, jcolors, LaCroixColoR, lisa, nationalparkcolors,
 #' NineteenEightyR, nord, ochRe, pals, Polychrome, MapPalettes, miscpalettes,
 #' rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, tidyquant, trekcolors,
-#' unikn, vapeplot, vapoRwave, werpals, wesanderson and yarrr.
+#' tvthemes, unikn, vapeplot, vapoRwave, werpals, wesanderson and yarrr.
 #'
 #' data.frame of palette names can be found here \code{\link{palettes_d_names}}.
 #'
-#' @format A list of 40 lists.
+#' @format A list of 41 lists.
 "palettes_d"
 
 #' Names of all fixed discrete palettes
@@ -28,11 +28,11 @@
 #' IslamicArt, jcolors, LaCroixColoR, lisa, nationalparkcolors,
 #' NineteenEightyR, nord, ochRe, pals, Polychrome, MapPalettes, miscpalettes,
 #' rcartocolor, RColorBrewer, Redmonder, RSkittleBrewer, tidyquant, trekcolors,
-#' unikn, vapeplot, vapoRwave, werpals, wesanderson and yarrr.
+#' tvthemes, unikn, vapeplot, vapoRwave, werpals, wesanderson and yarrr.
 #'
 #' Full list of palette can be found here \code{\link{palettes_d}}.
 #'
-#' @format A data.frame of 1334 observations with 4 variables:
+#' @format A data.frame of 1361 observations with 4 variables:
 #' \describe{
 #'   \item{package}{character, name of package}
 #'   \item{palette}{character, name of palette}