+++ data-raw/packages.R
@@ -48,7 +48,8 @@ paletteer_packages <- tibble::tribble(
   "calecopal", "an-bui/calecopal", "0.1.0", FALSE, NA,
   "colorblindr", "clauswilke/colorblindr", "0.1.0", FALSE, NA,
   "colRoz", "jacintak/colRoz", "0.2.2", FALSE, NA,
-  "unikn", "hneth/unikn", "0.2.0.9003", TRUE, "0.2.0"
+  "unikn", "hneth/unikn", "0.2.0.9003", TRUE, "0.2.0",
+  "tvthemes", "Ryo-N7/tvthemes", "1.1.0", TRUE, "1.1.0"
 ) %>%
   arrange(Name)